pyproject.toml
[tool.poetry]
name = "blog-python"
version = "0.1.0"
description = "Blog with Python and Poetry"
authors = ["Marcos V. Leal <marcosleal.prd@gmail.com>"]
maintainers = ["Marcos V. Leal <marcosleal.prd@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/marcosleal-prd/blog-python-poetry"
repository = "https://github.com/marcosleal-prd/blog-python-poetry"
keywords = ["Python", "Poetry", "Django", "Docker"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
psycopg2-binary = "^2.8.4"
django-filter = "^2.2.0"
drf-extensions = "^0.6.0"
python-json-logger = "^0.1.11"
request-id-django-log = "^0.1.1"
djangorestframework = "^3.11.0"
django-model-utils = "^4.0.0"
boto3 = "^1.12.39"
drf-link-navigation-pagination = "^0.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
isort = "^4.3.21"
pylint = "^2.4.4"
pylint-django = "^2.0.15"
pycodestyle = "^2.5.0"
flake8 = "^3.7.9"
mypy = "^0.770"
pytest-django = "^3.9.0"
pytest-env = "^0.6.2"
pytest-mock = "^3.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
force_single_line = true
line_length = 79
skip_glob = "**/migrations/*.py"
use_parentheses = true
multi_line_output = 5
include_trailing_comma = true

lines_between_types = 1

known_django = "django"
known_blog = "blog_python"
known_third_party = "rest_framework,_pytest,request_id_django_log,pytest,requests"
sections = "FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER,BLOG"

[tool.black]
line-length = 79
target_version = ["py37"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
)
'''
